Moved some Stripe folders to keep things tidier as we do more events
