+++ apps/integrations/src/integrations/stripe/webhooks/events/checkoutSession.examples.ts
